requirements.txt
streamlit==1.22.0
mysql-connector-python==8.0.33
Pillow==9.5.0